refactor: move vite-project to root dir
