+++ .storybook/preview-head.html
@@ -1,1 +1,4 @@
-<link rel="stylesheet" href="assets/styles/salesforce-lightning-design-system.min.css">
+<link rel="stylesheet" href="/assets/styles/salesforce-lightning-design-system.min.css">
+<script>
+  window.global = window;
+</script>
