+++ tsconfig.json
@@ -1,11 +1,21 @@
 {
   "compilerOptions": {
-    "target": "es5",
-    "module": "ES2020",
-    "lib": ["es2015", "dom"],
-    "jsx": "react",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "esModuleInterop": true
-  }
+    "target": "ESNext",
+    "useDefineForClassFields": true,
+    "lib": ["DOM", "DOM.Iterable", "ESNext"],
+    "allowJs": false,
+    "skipLibCheck": true,
+    "esModuleInterop": false,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "module": "ESNext",
+    "moduleResolution": "Node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "jsx": "react-jsx"
+  },
+  "include": ["src"],
+  "references": [{ "path": "./tsconfig.node.json" }]
 }
