--- vite-project/.storybook/main.js
@@ -1,12 +1,0 @@
-module.exports = {
-  stories: ["../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  staticDirs: ["../public"],
-  addons: [],
-  framework: "@storybook/react",
-  core: {
-    builder: "@storybook/builder-vite",
-  },
-  features: {
-    storyStoreV7: true,
-  },
-};
