--- vite-project/.storybook/preview-head.html
@@ -1,4 +1,0 @@
-<link rel="stylesheet" href="/assets/styles/salesforce-lightning-design-system.min.css">
-<script>
-  window.global = window;
-</script>
