--- vite-project/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "copy-title-and-url-as-markdown",
-  "version": "1.10.0",
-  "description": "Quickly copy the title & url of current tab as Markdown style (`[title](url)`)",
-  "private": true,
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview",
-    "test": "vitest",
-    "postinstall": "cp -r node_modules/@salesforce-ux/design-system/assets public",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
-    "zip": "bestzip dist.zip dist/*",
-    "lint": "eslint --ext .ts,.tsx src/**",
-    "release": "run-s lint build && standard-version",
-    "release:dry-run": "standard-version --dry-run"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zaki-yama/copy-title-and-url-as-markdown.git"
-  },
-  "keywords": [],
-  "author": "Shingo Yamazaki <shingoyamazaki00@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/zaki-yama/copy-title-and-url-as-markdown/issues"
-  },
-  "homepage": "https://github.com/zaki-yama/copy-title-and-url-as-markdown#readme",
-  "dependencies": {
-    "@salesforce-ux/design-system": "^2.18.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-lightning-design-system": "^5.2.4"
-  },
-  "devDependencies": {
-    "@crxjs/vite-plugin": "^1.0.3",
-    "@cybozu/eslint-config": "^17.0.1",
-    "@storybook/addon-essentials": "^6.5.9",
-    "@storybook/builder-vite": "^0.1.36",
-    "@storybook/react": "^6.5.9",
-    "@storybook/testing-library": "^0.0.13",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@vitejs/plugin-react": "^1.3.0",
-    "bestzip": "^2.2.1",
-    "eslint": "^8.18.0",
-    "typescript": "^4.6.3",
-    "vite": "^2.9.7",
-    "vitest": "^0.14.2"
-  }
-}
