--- vite-project/src/options/Options.css
@@ -1,14 +1,0 @@
-html {
-  background: #f3f2f2;
-}
-
-.optionsContainer {
-  width: 300px;
-  margin: 30px auto 0;
-}
-
-.toast {
-  width: 100%;
-  min-width: initial;
-  margin: 0 auto;
-}
