--- vite-project/src/options/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { Options } from "./Options";
-
-ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
-    <Options />
-  </React.StrictMode>
-);
