--- vite-project/src/popup/Popup.css
@@ -1,13 +1,0 @@
-html {
-  background: white;
-}
-
-.popupContainer {
-  width: 400px;
-  padding: 10px;
-}
-
-#dummy {
-  position: absolute;
-  top: -9999px;
-}
